13D006
This update resolves some performance and storage issues with the
statistics package introduced in the 13D005 release of Noteworthy.
People using the statistics package will see their old table orphaned
as the new ones will not use the older structure. Other advantages to
this implementation include the elimination for the need to use an
up-to-date .ini for PHP's get_broswer function, which would likely
return empty strings on shared hosts.